refactor: put some "static" styles to the separate css file
