+++ src/app/train/page.css
@@ -1,0 +1,8 @@
+.App {
+    text-align: center;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    height: 100vh;
+    transition: background 5s ease;
+}
